backend/cloud_functions/stream_processor/requirements.txt
functions-framework==3.8.2
confluent-kafka==2.6.1
google-cloud-aiplatform==1.75.0
